refactor: streamline service and repository structure, enhance context handling, and improve documentation
