+++ apps/api/src/api/v1/modules/categories/category.repository.ts
@@ -1,7 +1,7 @@
-import prisma, { Prisma } from '@poveroh/prisma'
 import type { CategoryData, CategoryFilters, CreateCategoryRequest, UpdateCategoryRequest } from '@poveroh/types'
-import { buildWhere } from '@/src/helpers/filter.helper'
-import { CATEGORY_TEMPLATE } from '@/src/api/v1/content/template/category'
+import { CATEGORY_TEMPLATE } from '@/v1/content/template/category'
+import prisma, { Prisma } from '@poveroh/prisma'
+import { buildWhere } from '@/helpers'
 
 const subcategorySelect = {
     id: true,
@@ -32,11 +32,11 @@ const categoryWithSubcategoriesSelect = {
 
 export class CategoryRepository {
     /**
-     * Creates a new category in the database, associating it with the specified user and using the provided payload for the category details. The method takes in the user ID to ensure that the category is linked to the correct user, and it uses the category ID and payload to populate the necessary fields in the database. The created category is then returned as a CategoryData object, which includes all relevant information about the category that was just created. This method is essential for allowing users to add new categories to their account, enabling them to organize their data effectively within the application.
-     * @param userId The ID of the user who is creating the category. This is used to associate the category with the correct user in the database, ensuring that each user's categories are kept separate and secure.
-     * @param id The unique identifier for the category being created. This ID is typically generated by the client or the server and is used to reference the category in future operations such as updates or deletions.
-     * @param payload An object containing the details of the category to be created, such as the title, icon, color, and any other relevant information that defines the category. This payload is used to populate the fields of the category in the database when it is created.
-     * @returns A promise that resolves to a CategoryData object representing the newly created category, including all relevant information such as the category's ID, title, icon, color, and timestamps for when it was created and last updated. This allows the client to have immediate access to the details of the category that was just created, which can be useful for updating the user interface or performing additional operations on the category right after its creation.
+     * Creates a new category in the database, associating it with the specified user and using the provided payload for the category details.
+     * @param userId The ID of the user who is creating the category.
+     * @param id The unique identifier for the category being created.
+     * @param payload An object containing the details of the category to be created, such as the title, icon, color, and any other relevant information that defines the category.
+     * @returns A promise that resolves to a CategoryData object representing the newly created category.
      */
     async create(userId: string, id: string, payload: CreateCategoryRequest): Promise<CategoryData> {
         return (await prisma.category.create({
@@ -46,9 +46,9 @@ export class CategoryRepository {
     }
 
     /**
-     * Creates a set of default categories based on a predefined template, associating them with the specified user. This method is designed to provide users with a starting point for organizing their data by automatically generating a standard set of categories that they can use immediately after signing up or when they choose to reset their categories. The method takes in the user ID to ensure that the created categories are linked to the correct user in the database, and it uses a predefined template (CATEGORY_TEMPLATE) to populate the details of each category, such as the title, icon, color, and any associated subcategories. The created categories are returned as an array of CategoryData objects, allowing the client to have immediate access to the details of all the categories that were just created from the template.
-     * @param userId The ID of the user for whom the default categories are being created. This ensures that the categories are associated with the correct user in the database, allowing each user to have their own set of default categories without interference from other users' data.
-     * @returns A promise that resolves to an array of CategoryData objects, each representing a category that was created based on the predefined template. These objects include all relevant information about the categories, such as their IDs, titles, icons, colors, and timestamps for when they were created and last updated. This allows the client to have immediate access to the details of all the default categories that were generated for the user, which can be useful for populating the user interface or performing additional operations on these categories right after their creation.
+     * Creates a set of default categories based on a predefined template, associating them with the specified user.
+     * @param userId The ID of the user for whom the default categories are being created.
+     * @returns A promise that resolves to an array of CategoryData objects, each representing a category that was created based on the predefined template.
      */
     async createFromTemplate(userId: string): Promise<CategoryData[]> {
         return (await prisma.$transaction(
@@ -74,10 +74,10 @@ export class CategoryRepository {
     }
 
     /**
-     * Updates an existing category in the database, ensuring that the category belongs to the specified user and is not soft-deleted. The method takes in the user ID to verify ownership of the category, the category ID to identify which category to update, and a payload containing the new details for the category. If the category exists and belongs to the user, it will be updated with the new information provided in the payload. This method is essential for allowing users to modify their existing categories, enabling them to keep their data organized and up-to-date as their needs change over time.
-     * @param userId The ID of the user who owns the category being updated. This is used to ensure that the category being modified belongs to the correct user, preventing unauthorized updates to categories that do not belong to the user.
-     * @param id The unique identifier of the category that is being updated. This ID is used to locate the specific category in the database that needs to be modified with the new details provided in the payload.
-     * @param payload  An object containing the new details for the category, such as an updated title, icon, color, or any other relevant information that defines the category. This payload is used to update the fields of the category in the database when the update operation is performed. The method does not return any value, but it ensures that the specified category is updated with the new information if it exists and belongs to the user.
+     * Updates an existing category in the database, ensuring that the category belongs to the specified user and is not soft-deleted.
+     * @param userId The ID of the user who owns the category being updated.
+     * @param id The unique identifier of the category that is being updated.
+     * @param payload  An object containing the new details for the category.
      */
     async update(userId: string, id: string, payload: UpdateCategoryRequest): Promise<void> {
         await prisma.category.update({
@@ -87,10 +87,10 @@ export class CategoryRepository {
     }
 
     /**
-     * Soft-deletes a category and its child subcategories in the same transaction. This method ensures that when a category is deleted, all associated subcategories are also marked as deleted, maintaining data consistency and preventing orphaned subcategories. The method takes in the user ID to verify ownership of the category, the category ID to identify which category to delete, and a timestamp indicating when the deletion occurred.
-     * @param userId The ID of the user who owns the category being deleted. This is used to ensure that the category being deleted belongs to the correct user, preventing unauthorized deletions of categories that do not belong to the user.
-     * @param id The unique identifier of the category that is being deleted. This ID is used to locate the specific category in the database that needs to be marked as deleted.
-     * @param deletedAt The timestamp indicating when the deletion occurred. This is used to mark the category and its subcategories as deleted in the database.
+     * Soft-deletes a category and its child subcategories in the same transaction.
+     * @param userId The ID of the user who owns the category being deleted.
+     * @param id The unique identifier of the category that is being deleted.
+     * @param deletedAt The timestamp indicating when the deletion occurred.
      */
     async softDelete(userId: string, id: string, deletedAt: Date): Promise<void> {
         await prisma.$transaction([
@@ -106,9 +106,9 @@ export class CategoryRepository {
     }
 
     /**
-     * Soft-deletes all categories owned by the specified user, along with their child subcategories, in the same transaction. This method is designed to allow users to delete all of their categories at once while ensuring that all associated subcategories are also marked as deleted, maintaining data consistency and preventing orphaned subcategories. The method takes in the user ID to identify which user's categories should be deleted and a timestamp indicating when the deletion occurred.
-     * @param userId The ID of the user whose categories are being deleted. This is used to ensure that only the categories belonging to the specified user are marked as deleted, preventing unauthorized deletions of other users' categories.
-     * @param deletedAt The timestamp indicating when the deletion occurred. This is used to mark all of the user's categories and their subcategories as deleted in the database.
+     * Soft-deletes all categories owned by the specified user, along with their child subcategories, in the same transaction.
+     * @param userId The ID of the user whose categories are being deleted.
+     * @param deletedAt The timestamp indicating when the deletion occurred.
      */
     async softDeleteAll(userId: string, deletedAt: Date): Promise<void> {
         await prisma.$transaction([
@@ -124,10 +124,10 @@ export class CategoryRepository {
     }
 
     /**
-     * Finds a category by its user-scoped ID, ensuring that the category belongs to the specified user and is not soft-deleted. This method is essential for retrieving the details of a specific category while enforcing ownership and visibility constraints. It takes in the user ID to verify that the category belongs to the correct user, and the category ID to identify which category to retrieve. If the category exists and meets the criteria, it returns a CategoryData object containing all relevant information about the category; otherwise, it returns null.
-     * @param userId The ID of the user who owns the category being retrieved. This is used to ensure that the category being accessed belongs to the correct user, preventing unauthorized access to categories that do not belong to the user.
-     * @param id The unique identifier of the category that is being retrieved. This ID is used to locate the specific category in the database that needs to be accessed.
-     * @returns A promise that resolves to a CategoryData object representing the category that was found, including all relevant information such as the category's ID, title, icon, color, and timestamps for when it was created and last updated. If no category is found that matches the criteria (i.e., it does not belong to the user or is soft-deleted), the method returns null. This allows the client to handle cases where the requested category does not exist or is not accessible due to ownership or deletion status.
+     * Finds a category by its user-scoped ID, ensuring that the category belongs to the specified user and is not soft-deleted.
+     * @param userId The ID of the user who owns the category being retrieved.
+     * @param id The unique identifier of the category that is being retrieved.
+     * @returns A promise that resolves to a CategoryData object representing the category that was found, or null if no category is found.
      */
     async findById(userId: string, id: string): Promise<CategoryData | null> {
         return (await prisma.category.findFirst({
@@ -137,12 +137,12 @@ export class CategoryRepository {
     }
 
     /**
-     * Finds categories based on the specified filters, ensuring that they belong to the specified user and are not soft-deleted. This method allows for flexible retrieval of categories by applying various filters while enforcing ownership and visibility constraints. It takes in the user ID to verify that the categories belong to the correct user, a set of filters to narrow down the search results, and pagination parameters (skip and take) to control the number of results returned. The method returns an array of CategoryData objects representing the categories that match the criteria, allowing the client to display or manipulate these categories as needed.
-     * @param userId The ID of the user who owns the categories being retrieved. This is used to ensure that only categories belonging to the specified user are returned, preventing unauthorized access to other users' categories.
-     * @param filters An object containing various filters that can be applied to narrow down the search results. These filters may include properties such as title, icon, color, or any other relevant attributes that can be used to identify specific categories. The method will use these filters to construct a query that retrieves only the categories that match the specified criteria while also ensuring that they are not soft-deleted and belong to the correct user.
-     * @param skip The number of categories to skip in the result set, used for pagination purposes. This allows the client to control which subset of the results is returned, enabling features such as infinite scrolling or page-based navigation through the categories.
-     * @param take The number of categories to return in the result set, used for pagination purposes. This allows the client to limit the number of categories returned in a single response, which can improve performance and user experience when dealing with large sets of categories. By specifying both skip and take, the client can effectively manage the retrieval of categories in a way that suits their needs while still enforcing ownership and visibility constraints.
-     * @returns A promise that resolves to an array of CategoryData objects representing the categories that were found based on the specified filters, ownership, and visibility criteria. Each CategoryData object includes all relevant information about the category, such as its ID, title, icon, color, and timestamps for when it was created and last updated. If no categories are found that match the criteria, the method returns an empty array. This allows the client to handle cases where no categories meet the specified filters or ownership requirements gracefully.
+     * Finds categories based on the specified filters, ensuring that they belong to the specified user and are not soft-deleted.
+     * @param userId The ID of the user who owns the categories being retrieved.
+     * @param filters An object containing various filters that can be applied to narrow down the search results.
+     * @param skip The number of categories to skip in the result set, used for pagination purposes.
+     * @param take The number of categories to return in the result set, used for pagination purposes.
+     * @returns A promise that resolves to an array of CategoryData objects representing the categories that were found based on the specified filters, ownership, and visibility criteria.
      */
     async findMany(userId: string, filters: CategoryFilters, skip: number, take: number): Promise<CategoryData[]> {
         const where = buildWhere({ ...filters, deletedAt: null, userId }, ['title'])
