+++ apps/api/src/api/v1/modules/dashboard/dashboard.controller.ts
@@ -1,15 +1,15 @@
 import { Request, Response } from 'express'
 import { GetDashboardLayout, UpdateDashboardLayoutRequest } from '@poveroh/types'
-import { BadRequestError, ResponseHelper } from '@/src/utils'
-import { DashboardService } from '@/src/api/v1/modules/dashboard/dashboard.service'
+import { BadRequestError, ResponseHelper } from '@/utils'
+import { DashboardService } from '@/v1/modules/dashboard/dashboard.service'
 
 export class DashboardController {
+    private static readonly dashboardService = new DashboardService()
+
     // GET /
     static async getDashboard(req: Request, res: Response) {
         try {
-            const dashboardService = new DashboardService()
-
-            const dashboardLayout = await dashboardService.getDashboardLayout()
+            const dashboardLayout = await this.dashboardService.getDashboardLayout()
             return ResponseHelper.success<GetDashboardLayout>(res, dashboardLayout)
         } catch (error) {
             return ResponseHelper.handleError(res, error)
@@ -25,8 +25,7 @@ export class DashboardController {
                 throw new BadRequestError('Layout not provided')
             }
 
-            const dashboardService = new DashboardService()
-            await dashboardService.saveDashboardLayout(payload)
+            await this.dashboardService.saveDashboardLayout(payload)
 
             return ResponseHelper.success(res)
         } catch (error) {
